requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
# passlib 1.7.4's bcrypt handler expects bcrypt<4 (bcrypt 4+ removed __about__).
# Also, we use bcrypt_sha256 (via passlib) to avoid bcrypt's 72-byte input limit.
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.34
httpx==0.27.2
python-multipart==0.0.9
